Crud final

+++ routes/web.php
@@ -25,7 +25,9 @@ Route::get('/empresas/{id}', [BusinessController::class,'show'])->name('empresa.
 Route::post('/user/register', [UserController::class,'store'])->name("user.register");
 Route::put('/user/{id}/update', [UserController::class,'update'])->name("user.update");
 Route::get('/user/{id}/dashboard', [UserController::class,'show'])->name("user.dashboard")->middleware('auth');
-Route::put('/usuario/update/{id}',[UserController::class,'update'])->name('usuario.update');
+
+///CRUD DE TODOS LOS USUARIOS-ADMIN
+Route::put('/usuario/update/{id}',[UserController::class,'updateAdmin'])->name('usuario.update');
 Route::delete('/usuario/{id}', [UserController::class,'destroy'])->name("usuario.delete");
 Route::put('/usuario/inactive/{id}', [UserController::class,'inactive'])->name("usuario.inactive");
 Route::put('/usuario/active/{id}', [UserController::class,'active'])->name("usuario.active");
@@ -43,18 +45,29 @@ Route::put('/social/{id}/update',[SocialController::class,'update'])->name('soci
 
 Route::put('/business/update/{id}', [BusinessController::class,'update'])->name('bussines.update');
 Route::delete('/business/{id}', [BusinessController::class,'destroy'])->name('bussines.delete');
-Route::put('/empresa/update/{id}',[BusinessController::class,'update'])->name('empresa.update');
-Route::delete('/empresa/{id}', [BusinessController::class,'destroy'])->name("empresa.delete");
+
+////RUTAS DE CRUD DE EMPRESAS ADMIN
+Route::put('/empresa/update/{id}',[BusinessController::class,'updateAdmin'])->name('empresa.update');
+Route::delete('/empresa/{id}', [BusinessController::class,'destroyAdmin'])->name("empresa.delete");
 Route::put('/empresa/inactive/{id}', [BusinessController::class,'inactive'])->name("empresa.inactive");
 Route::put('/empresa/active/{id}', [BusinessController::class,'active'])->name("empresa.active");
 
 
 Route::get('/business/register/offer/{id}',[OfferController::class,'create'])->name('offer.create');
 Route::post('/business/register/offer',[OfferController::class,'store'])->name('offer.register');
-Route::put('/offer/update/{id}',[OfferController::class,'update'])->name('offer.update');
-Route::delete('/offer/{id}', [OfferController::class,'destroy'])->name("offer.delete");
-Route::put('/offer/inactive/{id}', [OfferController::class,'inactive'])->name("offer.inactive");
-Route::put('/offer/active/{id}', [OfferController::class,'active'])->name("offer.active");
+
+///CRUD DE OFERTAS DE USUARO 
+Route::put('/offer/admin/update/{id}',[OfferController::class,'update'])->name('offer.update');
+Route::delete('/offer/admin/{id}', [OfferController::class,'destroy'])->name("offer.delete");
+Route::put('/offer/admin/inactive/{id}', [OfferController::class,'inactive'])->name("offer.inactive");
+Route::put('/offer/admin/active/{id}', [OfferController::class,'active'])->name("offer.active");
+
+//CRUD DE OFERTAS DE ADMINISTRADOR
+Route::put('/offer/update/{id}',[OfferController::class,'updateAdmin'])->name('ofertas.update');
+Route::delete('/offer/{id}', [OfferController::class,'destroyAdmin'])->name("ofertas.delete");
+Route::put('/offer/inactive/{id}', [OfferController::class,'inactiveAdmin'])->name("ofertas.inactive");
+Route::put('/offer/active/{id}', [OfferController::class,'activeAdmin'])->name("ofertas.active");
+
 
 Route::post('/ciudad/register',[CityController::class,'store'])->name('ciudad.register');
 Route::put('/ciudad/update/{id}',[CityController::class,'update'])->name('ciudad.update');
